Add config.json onboarding flow for workspace setup scripts
- Show toast notification when workspace is created without setup commands
- Add modal to configure .superset/config.json with setup/teardown scripts
- Create reusable OpenInButton component for opening files in various apps
- Add /scripts documentation page to website explaining config.json usage
- Migrate from setup.json to config.json naming convention
- Track toast dismissal per-project to avoid repeated prompts
