+++ apps/desktop/src/renderer/screens/main/components/StartView/index.tsx
@@ -1,4 +1,3 @@
-import { toast } from "@superset/ui/sonner";
 import { Tooltip, TooltipContent, TooltipTrigger } from "@superset/ui/tooltip";
 import { ChevronUp, FolderGit, FolderOpen, X } from "lucide-react";
 import { useState } from "react";
@@ -6,7 +5,6 @@ import { HiExclamationTriangle } from "react-icons/hi2";
 import { trpc } from "renderer/lib/trpc";
 import { useOpenNew } from "renderer/react-query/projects";
 import { useCreateWorkspace } from "renderer/react-query/workspaces";
-import { useOpenConfigModal } from "renderer/stores/config-modal";
 import { ActionCard } from "./ActionCard";
 import { CloneRepoDialog } from "./CloneRepoDialog";
 import { StartTopBar } from "./StartTopBar";
@@ -56,42 +54,17 @@ export function StartView() {
 	const { data: homeDir } = trpc.window.getHomeDir.useQuery();
 	const openNew = useOpenNew();
 	const createWorkspace = useCreateWorkspace();
-	const openConfigModal = useOpenConfigModal();
-	const dismissConfigToast = trpc.config.dismissConfigToast.useMutation();
 	const [error, setError] = useState<string | null>(null);
 	const [isCloneDialogOpen, setIsCloneDialogOpen] = useState(false);
 	const [showAllProjects, setShowAllProjects] = useState(false);
 	const [visibleCount, setVisibleCount] = useState(50);
-
-	const showConfigToastIfNeeded = (data: {
-		initialCommands: string[] | null;
-		projectId: string;
-	}) => {
-		if (!data.initialCommands || data.initialCommands.length === 0) {
-			toast.info("No setup script configured", {
-				description: "Automate workspace setup with a config.json file",
-				action: {
-					label: "Configure",
-					onClick: () => openConfigModal(data.projectId),
-				},
-				onDismiss: () => {
-					dismissConfigToast.mutate({ projectId: data.projectId });
-				},
-			});
-		}
-	};
 
 	const handleOpenProject = () => {
 		setError(null);
 		openNew.mutate(undefined, {
 			onSuccess: (result) => {
 				if (!result.canceled && result.project) {
-					createWorkspace.mutate(
-						{ projectId: result.project.id },
-						{
-							onSuccess: showConfigToastIfNeeded,
-						},
-					);
+					createWorkspace.mutate({ projectId: result.project.id });
 				}
 			},
 			onError: (err) => {
@@ -105,7 +78,6 @@ export function StartView() {
 		createWorkspace.mutate(
 			{ projectId },
 			{
-				onSuccess: showConfigToastIfNeeded,
 				onError: (err) => {
 					setError(err.message || "Failed to create workspace");
 				},
